removing working at first but not second time

+++ AirportService/CityService.cs
@@ -46,12 +46,13 @@ namespace AirportService
         public void Remove(Guid id)
         {
             var city = _airplaneContext.Cities.FirstOrDefault(c=> c.idCity==id);
+            if(city!=null)
             _airplaneContext.Cities.Remove(city);
+
             var flight = _airplaneContext.Flights.Where(f => (f.idCityArrival == id || f.idCityDeparture == id));
+            if(flight!=null)
             _airplaneContext.Flights.RemoveRange(flight);
             _airplaneContext.SaveChanges();
-
-
         }
     }
 }
